fraudDetector/requirements.txt
Flask==1.1.2
Flask-BasicAuth==0.2.0
Flask-Cors==3.0.10
Flask-HTTPAuth==4.2.0
gunicorn==20.1.0
kafka-python-2.0.2
PyMySQL==1.0.2
requests==2.25.1
